requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=6.0

# GUI and plotting
PyQt6>=6.4.0
pyqtgraph>=0.13.0

# Data logging
h5py>=3.7.0

# MCC DAQ (Windows only)
# Install from: https://github.com/mccdaq/mcculw
# Or: pip install mcculw
mcculw; platform_system=="Windows"
